Register the <Plug> commands even when default key mappings are disabled
The check for `bookmark_no_default_key_mappings` was too aggressive, not
registering the <Plug> aliases at all. This meant that the instructions for
overriding shortcuts from the README did not work.

This change moves the check of that global to the place where the default
shortcuts are actually added after the <Plug> alias is regsitered.